requirements.txt
asgiref==3.2.10
boto3==1.17.35
botocore==1.20.35
Brotli==1.0.9
dj-database-url==0.5.0
Django==3.1.1
django-appconf==1.0.4
django-compression-middleware==0.4.1
django-compressor==2.4
django-endless-pagination==2.0
django-environ==0.4.5
django-filer==2.0.2
django-filter==2.4.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-next-prev==1.0.1
django-polymorphic==3.0.0
django-resized==0.3.11
django-ses==2.0.0
django-storages==1.11.1
easy-thumbnails==2.7.1
future==0.18.2
gunicorn==20.0.4
jmespath==0.10.0
jsonfield==3.1.0
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.3.1
tzlocal==2.1
Unidecode==1.1.2
urllib3==1.26.4
whitenoise==5.2.0
zstandard==0.15.2
